+++ @xen-orchestra/lite/src/components/TitleBar.vue
@@ -1,10 +1,8 @@
 <template>
-  <div class="container">
-    <div class="title-bar">
-      <FontAwesomeIcon :icon="icon" class="icon" />
-      <div class="title">
-        <slot />
-      </div>
+  <div class="title-bar">
+    <FontAwesomeIcon :icon="icon" class="icon" />
+    <div class="title">
+      <slot />
     </div>
     <div class="actions">
       <slot name="actions" />
@@ -21,19 +19,17 @@ defineProps<{
 </script>
 
 <style lang="postcss" scoped>
-.container {
+.actions {
+  margin-left: auto;
+}
+
+.title-bar {
   display: flex;
   align-items: center;
   height: 8rem;
   padding: 0 2rem;
   border-bottom: 1px solid var(--color-blue-scale-400);
   background-color: var(--background-color-primary);
-  justify-content: space-between;
-}
-
-.title-bar {
-  display: flex;
-  align-items: center;
   gap: 1.5rem;
 }
