+++ @xen-orchestra/lite/src/locales/fr.json
@@ -12,7 +12,7 @@
   "back-pool-dashboard": "Revenez au tableau de bord de votre pool",
   "cancel": "Annuler",
   "change-power-state": "Changer l'état d'alimentation",
-  "change-state": "Changer d'état",
+  "change-state": "Changer l'état",
   "community": "Communauté",
   "community-name": "Communauté {name}",
   "copy": "Copier",
