+++ @xen-orchestra/lite/src/views/vm/VmRootView.vue
@@ -11,7 +11,7 @@
 
 <script lang="ts" setup>
 import ObjectNotFoundWrapper from "@/components/ObjectNotFoundWrapper.vue";
-import VmHeader from "@/views/vm/VmHeader.vue";
+import VmHeader from "@/components/vm/VmHeader.vue";
 import { watchEffect } from "vue";
 import { useRoute } from "vue-router";
 import { useUiStore } from "@/stores/ui.store";
